fix travis file to use appropriate jar file
The two commits ago was more on track than the last commit. The
ScriptManager jar is indeed in the build/libs/* directory but I failed
to use the write name for the jarfile. In this commit, I made the jar
file name generic enough that furture changes to the name like version
changes should work with this .travis.yml